Improved image upload (#58)
Implementation of new upload strategy. 
Basic concept: Hold a list of items to upload which is cut off if it
gets to long.
In each update cycle the first BS items are uploaded.
There is a priority queue with images that are not dropped and always
uploaded fist.

Fixes an error due to an unclosed clientsession where we forgot to await
the response.

BREAKING CHANGE:
The abstract method `evaluate` which needs to be implemented by
detectors now provides the raw jpg image bytes instead of a
semi-converted np.array. This means if the detector uses cv2, reading
the file can be done with `cv_image = cv2.imdecode(np.frombuffer(image,
np.uint8), cv2.IMREAD_COLOR)`. Alternatively PIL could be used as well.

---------

Co-authored-by: Niklas Neugebauer <niklas@zauberzeug.com>