fix release of detection lock (#66)
In the detector we introduced a lock to make sure no inference overlaps
with another one. However, when there was an exception in the inference
(e.g. model not initialized) the lock was never released.

This PR also fixes that we missed to set metadata for the batch evaluate
function.